+++ src/input_streams.rs
@@ -4,9 +4,9 @@ use bevy::ecs::prelude::{Event, Events, ResMut, World};
 use bevy::ecs::system::SystemState;
 use bevy::input::{
     gamepad::{Gamepad, GamepadAxis, GamepadButton, GamepadEvent, Gamepads},
-    keyboard::{KeyCode, KeyboardInput, ScanCode},
+    keyboard::{KeyCode, KeyboardInput},
     mouse::{MouseButton, MouseButtonInput, MouseMotion, MouseWheel},
-    Axis, Input,
+    Axis, ButtonInput,
 };
 use bevy::math::Vec2;
 use bevy::utils::HashSet;
@@ -19,25 +19,23 @@ use crate::buttonlike::{MouseMotionDirection, MouseWheelDirection};
 use crate::prelude::DualAxis;
 use crate::user_input::{InputKind, UserInput};
 
-/// A collection of [`Input`] structs, which can be used to update an [`InputMap`](crate::input_map::InputMap).
+/// A collection of [`ButtonInput`] structs, which can be used to update an [`InputMap`](crate::input_map::InputMap).
 ///
 /// These are typically collected via a system from the [`World`] as resources.
 #[derive(Debug, Clone)]
 pub struct InputStreams<'a> {
-    /// A [`GamepadButton`] [`Input`] stream
-    pub gamepad_buttons: &'a Input<GamepadButton>,
+    /// A [`GamepadButton`] [`Input`](ButtonInput) stream
+    pub gamepad_buttons: &'a ButtonInput<GamepadButton>,
     /// A [`GamepadButton`] [`Axis`] stream
     pub gamepad_button_axes: &'a Axis<GamepadButton>,
     /// A [`GamepadAxis`] [`Axis`] stream
     pub gamepad_axes: &'a Axis<GamepadAxis>,
     /// A list of registered gamepads
     pub gamepads: &'a Gamepads,
-    /// A [`KeyCode`] [`Input`] stream
-    pub keycodes: Option<&'a Input<KeyCode>>,
-    /// A [`ScanCode`] [`Input`] stream
-    pub scan_codes: Option<&'a Input<ScanCode>>,
-    /// A [`MouseButton`] [`Input`] stream
-    pub mouse_buttons: Option<&'a Input<MouseButton>>,
+    /// A [`KeyCode`] [`ButtonInput`] stream
+    pub keycodes: Option<&'a ButtonInput<KeyCode>>,
+    /// A [`MouseButton`] [`Input`](ButtonInput) stream
+    pub mouse_buttons: Option<&'a ButtonInput<MouseButton>>,
     /// A [`MouseWheel`] event stream
     pub mouse_wheel: Option<Vec<MouseWheel>>,
     /// A [`MouseMotion`] event stream
@@ -50,13 +48,12 @@ pub struct InputStreams<'a> {
 impl<'a> InputStreams<'a> {
     /// Construct an [`InputStreams`] from a [`World`]
     pub fn from_world(world: &'a World, gamepad: Option<Gamepad>) -> Self {
-        let gamepad_buttons = world.resource::<Input<GamepadButton>>();
+        let gamepad_buttons = world.resource::<ButtonInput<GamepadButton>>();
         let gamepad_button_axes = world.resource::<Axis<GamepadButton>>();
         let gamepad_axes = world.resource::<Axis<GamepadAxis>>();
         let gamepads = world.resource::<Gamepads>();
-        let keycodes = world.get_resource::<Input<KeyCode>>();
-        let scan_codes = world.get_resource::<Input<ScanCode>>();
-        let mouse_buttons = world.get_resource::<Input<MouseButton>>();
+        let keycodes = world.get_resource::<ButtonInput<KeyCode>>();
+        let mouse_buttons = world.get_resource::<ButtonInput<MouseButton>>();
         let mouse_wheel = world.resource::<Events<MouseWheel>>();
         let mouse_motion = world.resource::<Events<MouseMotion>>();
 
@@ -69,7 +66,6 @@ impl<'a> InputStreams<'a> {
             gamepad_axes,
             gamepads,
             keycodes,
-            scan_codes,
             mouse_buttons,
             mouse_wheel: Some(mouse_wheel),
             mouse_motion,
@@ -120,11 +116,8 @@ impl<'a> InputStreams<'a> {
                         button_type,
                     })
                 }),
-            InputKind::Keyboard(keycode) => {
+            InputKind::PhysicalKey(keycode) => {
                 matches!(self.keycodes, Some(keycodes) if keycodes.pressed(keycode))
-            }
-            InputKind::KeyLocation(scan_code) => {
-                matches!(self.scan_codes, Some(scan_codes) if scan_codes.pressed(scan_code))
             }
             InputKind::Modifier(modifier) => {
                 let key_codes = modifier.key_codes();
@@ -378,14 +371,14 @@ fn collect_events_cloned<T: Event + Clone>(events: &Events<T>) -> Vec<T> {
     events.get_reader().read(events).cloned().collect()
 }
 
-/// A mutable collection of [`Input`] structs, which can be used for mocking user inputs.
+/// A mutable collection of [`ButtonInput`] structs, which can be used for mocking user inputs.
 ///
 /// These are typically collected via a system from the [`World`] as resources.
 // WARNING: If you update the fields of this type, you must also remember to update `InputMocking::reset_inputs`.
 #[derive(Debug)]
 pub struct MutableInputStreams<'a> {
-    /// A [`GamepadButton`] [`Input`] stream
-    pub gamepad_buttons: &'a mut Input<GamepadButton>,
+    /// A [`GamepadButton`] [`Input`](ButtonInput) stream
+    pub gamepad_buttons: &'a mut ButtonInput<GamepadButton>,
     /// A [`GamepadButton`] [`Axis`] stream
     pub gamepad_button_axes: &'a mut Axis<GamepadButton>,
     /// A [`GamepadAxis`] [`Axis`] stream
@@ -395,15 +388,13 @@ pub struct MutableInputStreams<'a> {
     /// Events used for mocking gamepad-related inputs
     pub gamepad_events: &'a mut Events<GamepadEvent>,
 
-    /// A [`KeyCode`] [`Input`] stream
-    pub keycodes: &'a mut Input<KeyCode>,
-    /// A [`ScanCode`] [`Input`] stream
-    pub scan_codes: &'a mut Input<ScanCode>,
+    /// A [`KeyCode`] [`ButtonInput`] stream
+    pub keycodes: &'a mut ButtonInput<KeyCode>,
     /// Events used for mocking keyboard-related inputs
     pub keyboard_events: &'a mut Events<KeyboardInput>,
 
-    /// A [`MouseButton`] [`Input`] stream
-    pub mouse_buttons: &'a mut Input<MouseButton>,
+    /// A [`MouseButton`] [`Input`](ButtonInput) stream
+    pub mouse_buttons: &'a mut ButtonInput<MouseButton>,
     /// Events used for mocking [`MouseButton`] inputs
     pub mouse_button_events: &'a mut Events<MouseButtonInput>,
     /// A [`MouseWheel`] event stream
@@ -419,15 +410,14 @@ impl<'a> MutableInputStreams<'a> {
     /// Construct a [`MutableInputStreams`] from the [`World`]
     pub fn from_world(world: &'a mut World, gamepad: Option<Gamepad>) -> Self {
         let mut input_system_state: SystemState<(
-            ResMut<Input<GamepadButton>>,
+            ResMut<ButtonInput<GamepadButton>>,
             ResMut<Axis<GamepadButton>>,
             ResMut<Axis<GamepadAxis>>,
             ResMut<Gamepads>,
             ResMut<Events<GamepadEvent>>,
-            ResMut<Input<KeyCode>>,
-            ResMut<Input<ScanCode>>,
+            ResMut<ButtonInput<KeyCode>>,
             ResMut<Events<KeyboardInput>>,
-            ResMut<Input<MouseButton>>,
+            ResMut<ButtonInput<MouseButton>>,
             ResMut<Events<MouseButtonInput>>,
             ResMut<Events<MouseWheel>>,
             ResMut<Events<MouseMotion>>,
@@ -440,7 +430,6 @@ impl<'a> MutableInputStreams<'a> {
             gamepads,
             gamepad_events,
             keycodes,
-            scan_codes,
             keyboard_events,
             mouse_buttons,
             mouse_button_events,
@@ -455,7 +444,6 @@ impl<'a> MutableInputStreams<'a> {
             gamepads: gamepads.into_inner(),
             gamepad_events: gamepad_events.into_inner(),
             keycodes: keycodes.into_inner(),
-            scan_codes: scan_codes.into_inner(),
             keyboard_events: keyboard_events.into_inner(),
             mouse_buttons: mouse_buttons.into_inner(),
             mouse_button_events: mouse_button_events.into_inner(),
@@ -485,7 +473,6 @@ impl<'a> From<MutableInputStreams<'a>> for InputStreams<'a> {
             gamepad_axes: mutable_streams.gamepad_axes,
             gamepads: mutable_streams.gamepads,
             keycodes: Some(mutable_streams.keycodes),
-            scan_codes: Some(mutable_streams.scan_codes),
             mouse_buttons: Some(mutable_streams.mouse_buttons),
             mouse_wheel: Some(collect_events_cloned(mutable_streams.mouse_wheel)),
             mouse_motion: collect_events_cloned(mutable_streams.mouse_motion),
@@ -502,7 +489,6 @@ impl<'a> From<&'a MutableInputStreams<'a>> for InputStreams<'a> {
             gamepad_axes: mutable_streams.gamepad_axes,
             gamepads: mutable_streams.gamepads,
             keycodes: Some(mutable_streams.keycodes),
-            scan_codes: Some(mutable_streams.scan_codes),
             mouse_buttons: Some(mutable_streams.mouse_buttons),
             mouse_wheel: Some(collect_events_cloned(mutable_streams.mouse_wheel)),
             mouse_motion: collect_events_cloned(mutable_streams.mouse_motion),